ddl update 优化

+++ tools/generator/src/main/kotlin/top/bettercode/summer/tools/generator/ddl/MysqlToDDL.kt
@@ -12,7 +12,9 @@ object MysqlToDDL : ToDDL() {
         out.appendLine("$commentPrefix ${database.url.substringBefore("?")}")
         out.appendLine("$commentPrefix use ${database.schema};")
         out.appendLine()
-        if (!database.offline) {
+        val notOnlyCollate = !database.extension.setting("onlyCollate", "false").toBoolean()
+        val notOnlyComment = !database.extension.setting("onlyComment", "false").toBoolean()
+        if (!database.offline && notOnlyComment) {
             //schema default collate change
             database.use {
                 val defaultCollate = getSchemaDefaultCollate()
@@ -32,7 +34,7 @@ object MysqlToDDL : ToDDL() {
             }
             val tableNames = tables.map { it.tableName }
             val oldTableNames = oldTables.map { it.tableName }
-            if (database.dropTablesWhenUpdate) (oldTableNames - tableNames.toSet()).filter { "api_token" != it }.forEach {
+            if (notOnlyCollate && notOnlyComment && database.dropTablesWhenUpdate) (oldTableNames - tableNames.toSet()).filter { "api_token" != it }.forEach {
                 out.appendLine("$commentPrefix DROP $it")
                 out.appendLine("DROP TABLE IF EXISTS $schema$quote$it$quote;")
                 out.appendLine()
@@ -41,74 +43,88 @@ object MysqlToDDL : ToDDL() {
             tables.forEach { table ->
                 val tableName = table.tableName
                 if (newTableNames.contains(tableName)) {
-                    appendTable(schema, table, out, database)
+                    if (notOnlyCollate && notOnlyComment)
+                        appendTable(schema, table, out, database)
                 } else {
                     val oldTable = oldTables.find { it.tableName == tableName }!!
                     if (oldTable != table) {
                         val lines = mutableListOf<String>()
-                        if (oldTable.remarks != table.remarks) lines.add("ALTER TABLE $schema$quote$tableName$quote COMMENT '${
-                            table.remarks.replace("\\", "\\\\")
-                        }';")
+                        if (notOnlyCollate) {
+                            if (oldTable.remarks != table.remarks) lines.add("ALTER TABLE $schema$quote$tableName$quote COMMENT '${
+                                table.remarks.replace("\\", "\\\\")
+                            }';")
+
+                            val oldColumns = oldTable.columns
+                            val columns = table.columns
+                            val oldPrimaryKeys = oldTable.primaryKeys.toMutableSet()
+                            val primaryKeys = table.primaryKeys.toSet()
 
 
-                        val oldColumns = oldTable.columns
-                        val columns = table.columns
-                        val oldPrimaryKeys = oldTable.primaryKeys.toMutableSet()
-                        val primaryKeys = table.primaryKeys.toSet()
-
-
-                        val oldColumnNames = oldColumns.map { it.columnName }
-                        val columnNames = columns.map { it.columnName }
-                        val dropColumnNames = oldColumnNames - columnNames.toSet()
-                        if (database.dropColumnsWhenUpdate) dropColumnNames.forEach {
-                            lines.add("ALTER TABLE $schema$quote$tableName$quote DROP COLUMN $quote$it$quote;")
-                            oldPrimaryKeys.removeIf { pk -> pk.columnName == it }
-                        }
-                        dropFk(schema, oldColumns, dropColumnNames, lines, tableName)
-                        val newColumnNames = columnNames - oldColumnNames.toSet()
-                        columns.forEach { column ->
-                            val columnName = column.columnName
-                            if (newColumnNames.contains(columnName)) {
-                                lines.add("ALTER TABLE $schema$quote$tableName$quote ADD COLUMN ${
-                                    columnDef(column, quote)
-                                } COMMENT '${column.remarks.replace("\\", "\\\\")}';")
-                                addFk(schema, column, lines, tableName, columnName)
-                            } else {
-                                val oldColumn = oldColumns.find { it.columnName == columnName }!!
-                                if (column != oldColumn) {
-                                    lines.add("ALTER TABLE $schema$quote$tableName$quote MODIFY ${
-                                        columnDef(column, quote)
-                                    } COMMENT '${column.remarks.replace("\\", "\\\\")}';")
-                                    updateFk(schema, column, oldColumn, lines, tableName)
+                            val oldColumnNames = oldColumns.map { it.columnName }
+                            val columnNames = columns.map { it.columnName }
+                            val dropColumnNames = oldColumnNames - columnNames.toSet()
+                            if (notOnlyComment) {
+                                if (database.dropColumnsWhenUpdate) dropColumnNames.forEach {
+                                    lines.add("ALTER TABLE $schema$quote$tableName$quote DROP COLUMN $quote$it$quote;")
+                                    oldPrimaryKeys.removeIf { pk -> pk.columnName == it }
+                                }
+                                dropFk(schema, oldColumns, dropColumnNames, lines, tableName)
+                            }
+                            val newColumnNames = columnNames - oldColumnNames.toSet()
+                            columns.forEach { column ->
+                                val columnName = column.columnName
+                                if (newColumnNames.contains(columnName)) {
+                                    if (notOnlyComment) {
+                                        lines.add("ALTER TABLE $schema$quote$tableName$quote ADD COLUMN ${
+                                            columnDef(column, quote)
+                                        } COMMENT '${column.remarks.replace("\\", "\\\\")}';")
+                                        addFk(schema, column, lines, tableName, columnName)
+                                    }
+                                } else {
+                                    val oldColumn = oldColumns.find { it.columnName == columnName }!!
+                                    if (column != oldColumn) {
+                                        if (notOnlyComment) {
+                                            lines.add("ALTER TABLE $schema$quote$tableName$quote MODIFY ${
+                                                columnDef(column, quote)
+                                            } COMMENT '${column.remarks.replace("\\", "\\\\")}';")
+                                            updateFk(schema, column, oldColumn, lines, tableName)
+                                        } else if (oldColumn.remarks != column.remarks) {
+                                            lines.add("ALTER TABLE $schema$quote$tableName$quote MODIFY ${
+                                                columnDef(oldColumn, quote)
+                                            } COMMENT '${column.remarks.replace("\\", "\\\\")}';")
+                                        }
+                                    }
                                 }
                             }
-                        }
 
-                        if (oldPrimaryKeys.size == 1 && primaryKeys.size == 1) {
-                            val oldPrimaryKey = oldPrimaryKeys.first()
-                            val primaryKey = primaryKeys.first()
-                            if (oldPrimaryKey != primaryKey) {
-                                lines.add("ALTER TABLE $schema$quote$tableName$quote CHANGE $quote${oldPrimaryKey.columnName}$quote ${
-                                    columnDef(primaryKey, quote)
-                                } COMMENT '${primaryKey.remarks.replace("\\", "\\\\")}';")
+                            if (notOnlyComment) {
+                                if (oldPrimaryKeys.size == 1 && primaryKeys.size == 1) {
+                                    val oldPrimaryKey = oldPrimaryKeys.first()
+                                    val primaryKey = primaryKeys.first()
+                                    if (oldPrimaryKey != primaryKey) {
+                                        lines.add("ALTER TABLE $schema$quote$tableName$quote CHANGE $quote${oldPrimaryKey.columnName}$quote ${
+                                            columnDef(primaryKey, quote)
+                                        } COMMENT '${primaryKey.remarks.replace("\\", "\\\\")}';")
+                                    }
+                                } else if (oldPrimaryKeys != primaryKeys) {
+                                    if (oldPrimaryKeys.isNotEmpty())
+                                        lines.add("ALTER TABLE $schema$quote$tableName$quote DROP PRIMARY KEY;")
+                                    if (table.primaryKeyNames.isNotEmpty()) {
+                                        lines.add("ALTER TABLE $schema$quote$tableName$quote ADD PRIMARY KEY(${table.primaryKeyNames.joinToString(",") { "$quote$it$quote" }});")
+                                    }
+                                }
+
+                                if (database.queryIndex) updateIndexes(schema, oldTable, table, lines, dropColumnNames)
+
+                                //engine change
+                                if (!oldTable.engine.equals(table.engine, true)) lines.add("ALTER TABLE $schema$quote$tableName$quote ENGINE ${table.engine};")
                             }
-                        } else if (oldPrimaryKeys != primaryKeys) {
-                            if (oldPrimaryKeys.isNotEmpty())
-                                lines.add("ALTER TABLE $schema$quote$tableName$quote DROP PRIMARY KEY;")
-                            if (table.primaryKeyNames.isNotEmpty()) {
-                                lines.add("ALTER TABLE $schema$quote$tableName$quote ADD PRIMARY KEY(${table.primaryKeyNames.joinToString(",") { "$quote$it$quote" }});")
-                            }
                         }
-
-
-                        if (database.queryIndex) updateIndexes(schema, oldTable, table, lines, dropColumnNames)
-
-                        //engine change
-                        if (!oldTable.engine.equals(table.engine, true)) lines.add("ALTER TABLE $schema$quote$tableName$quote ENGINE ${table.engine};")
-
-                        //charset change
-                        if (oldTable.charset != table.charset || oldTable.collate != table.collate) {
-                            lines.add("ALTER TABLE $schema$quote$tableName$quote CONVERT TO CHARACTER SET ${table.charset} COLLATE ${table.collate};")
+                        if (notOnlyComment) {
+                            //charset change
+                            if (oldTable.charset != table.charset || oldTable.collate != table.collate) {
+                                lines.add("ALTER TABLE $schema$quote$tableName$quote CONVERT TO CHARACTER SET ${table.charset} COLLATE ${table.collate};")
+                            }
                         }
                         //out change
                         if (lines.isNotEmpty()) {
